+++ src/Transposition.h
@@ -119,9 +119,9 @@ public:
     static const u08 MAX_HASH_BIT;
 
     // Minimum size for Transposition table in mega-byte
-    static const u64 MIN_TT_SIZE;
+    static const u32 MIN_TT_SIZE;
     // Maximum size for Transposition table in mega-byte
-    static const u64 MAX_TT_SIZE;
+    static const u32 MAX_TT_SIZE;
 
     static bool Clear_Hash;
